:sparkles: Full scrape working
