+++ package.json
@@ -47,7 +47,6 @@
     "@types/imagemin-gifsicle": "^5.2.0",
     "@types/imagemin-jpegtran": "^5.0.0",
     "@types/node": "^10.10.1",
-    "@types/node-fetch": "^2.1.4",
     "@types/redis": "^2.8.10",
     "@types/rimraf": "^2.0.2",
     "@types/semver": "^5.5.0",
@@ -71,7 +70,6 @@
     "mkdirp": "^0.5.1",
     "mocha": "^5.2.0",
     "moment": "^2.21.0",
-    "node-fetch": "^1.7.3",
     "parsoid": "^0.10.0",
     "redis": "^2.8.0",
     "rimraf": "^2.6.3",
