+++ src/util/const.ts
@@ -1,0 +1,3 @@
+
+export const MEDIA_REGEX = /^(.*\/)([^/]+)(\/)(\d+px-|)(.+?)(\.[A-Za-z0-9]{2,6}|)(\.[A-Za-z0-9]{2,6}|)$/;
+export const MIN_IMAGE_THRESHOLD_ARTICLELIST_PAGE = 10;
