+++ src/util/index.ts
@@ -3,3 +3,4 @@ export * from './misc';
 export * from './dump';
 export * from './mapLimit';
 export * from './articleRenderers';
+export * from './const';
